pyproject.toml
[tool.poetry]
name = "coss"
version = "0.1.0"
description = ""
authors = ["Mikael Göransson <github@mgor.se>"]

[tool.poetry.scripts]
coss = "coss.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.1"
marshmallow-dataclass = "^8.5.3"
"ruamel.yaml" = "^0.17.19"
beautifulsoup4 = "^4.10.0"
Pillow = "^8.4.0"
python-telegram-bot = "^13.9"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
